WebMusicLibrary/catalog/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Album, Singer, Song

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'last_name', 'first_name']

class AlbumSerializer(serializers.ModelSerializer):
    class Meta:
        model = Album
        fields = ['id', 'title', 'release_date', 'name_singer', 'author']

class SingerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Singer
        fields = ['id', 'name_singer', 'author']

class SongSerializer(serializers.ModelSerializer):
    class Meta:
        model = Song
        fields = ['id', 'name_song', 'number', 'title', 'name_singer', 'author']
