requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset_normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
django_rest_swagger==2.2.0
Django==4.1.4
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi_codec==1.3.2
pip==22.3.1
pytz==2022.7
requests==2.28.1
rest-swagger==1.0.0
setuptools==56.0.0
simplejson==3.18.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wheel==0.38.4
whitenoise==6.3.0
